WIP: Moved dummy integration to its own root-level package
