+++ everpli_dummy/dummy_slider.py
@@ -1,17 +1,21 @@
 # Include standard modules
-# Include 3rd-party modules
+import time
 
+# Include 3rd-party modules
 # Include DPL modules
-from dpl.things import Player
+from dpl.things import Slider
 from dpl.integrations import ThingFactory, ThingRegistry
 from dpl.model.domain_id import TDomainId
-from . import DummyConnection
+
+from .dummy_connection import DummyConnection
 
 
-class DummyPlayer(Player):
+class DummySlider(Slider):
     """
-    A reference implementation of Player
+    A reference implementation of slider
     """
+    __SWITCH_DELAY = 1  # second
+
     def __init__(self, domain_id: TDomainId, con_instance: DummyConnection, con_params: dict, metadata: dict):
         """
         Constructor. Receives an instance of DummyConnection and a prefix to be printed
@@ -34,7 +38,7 @@ class DummyPlayer(Player):
         self._con_instance = con_instance
 
     @property
-    def state(self) -> Player.States:
+    def state(self) -> Slider.States:
         """
         Return a current state of the Thing
 
@@ -73,50 +77,58 @@ class DummyPlayer(Player):
         """
         self._is_enabled = True
 
-    def play(self, song_name: str = None) -> None:
+    def open(self) -> None:
         """
-        Starts playing and switches the object to the 'playing' state. Additional parameters
-        like track name or URL can be provided.
-
-        :param song_name: song name to be played (or URL, or ID, or playlist position, etc.)
-        :return: None
-        """
-        self._check_is_available()
-        self._con_instance.print(self._print_prefix, "Player is playing {0}".format(song_name))
-        self._state = self.States.playing
-
-    def stop(self) -> None:
-        """
-        Stops playing and switches the object to the 'stopped' state.
+        Switches an object to the 'opening' and then 'opened' state if its current state
+        is 'undefined', 'closed' or 'closing'. Command must be ignored otherwise.
 
         :return: None
         """
         self._check_is_available()
-        self._con_instance.print(self._print_prefix, "Player is stopped")
-        self._state = self.States.stopped
 
-    def pause(self) -> None:
+        if self._state == self.States.opening or self._state == self.States.opened:
+            pass
+        else:
+            self._con_instance.print(self._print_prefix, "Switch is opening...")
+            self._state = self.States.opening
+
+            time.sleep(self.__SWITCH_DELAY)
+
+            self._con_instance.print(self._print_prefix, "Switch is opened")
+            self._state = self.States.opened
+
+    def close(self) -> None:
         """
-        Pause playing and switches the object to the 'paused' state.
+        Switches an object to the 'closing' and then 'closed' state if its current state
+        is 'undefined', 'opened' or 'opening'. Command must be ignored otherwise.
 
         :return: None
         """
         self._check_is_available()
-        self._con_instance.print(self._print_prefix, "Player is paused")
-        self._state = self.States.paused
+
+        if self._state == self.States.closing or self._state == self.States.closed:
+            pass
+        else:
+            self._con_instance.print(self._print_prefix, "Switch is closing...")
+            self._state = self.States.closing
+
+            time.sleep(self.__SWITCH_DELAY)
+
+            self._con_instance.print(self._print_prefix, "Switch is closed")
+            self._state = self.States.closed
 
 
-class DummyPlayerFactory(ThingFactory):
+class DummySliderFactory(ThingFactory):
     """
-    DummyPlayerFactory is a class that is responsible for building of DummyPlayers
+    DummySliderFactory is a class that is responsible for building of DummySliders
     """
     @staticmethod
-    def build(*args, **kwargs) -> DummyPlayer:
-        return DummyPlayer(*args, **kwargs)
+    def build(*args, **kwargs) -> DummySlider:
+        return DummySlider(*args, **kwargs)
 
 
 ThingRegistry.register_factory(
     integration_name="dummy",
-    thing_type="player",
-    factory=DummyPlayerFactory()
+    thing_type="slider",
+    factory=DummySliderFactory()
 )
